requirements.txt
Flask>=2.1.2
Flask-SQLAlchemy>=2.5.1
Flask-Login>=0.6.2
Flask-Admin>=1.6.0
Flask-Migrate>=3.1.0
Werkzeug>=2.1.2
PyMySQL>=1.0.2
WTForms>=3.0.1
click>=8.1.3
itsdangerous>=2.1.2
Jinja2>=3.1.2
MarkupSafe>=2.1.1
SQLAlchemy>=1.4.39
alembic>=1.8.0
python-dotenv>=0.20.0
scipy>=1.5.0
pandas>=1.3.0
numpy>=1.20.0
gunicorn>=20.1.0
mysql-connector-python>=8.0.33

